move calculation funcs to utils
