+++ tests/time_tests/test_runner/conftest.py
@@ -23,10 +23,10 @@ import os
 import shutil
 import sys
 import tempfile
-from pathlib import Path
-
 import pytest
 import yaml
+
+from pathlib import Path
 from jsonschema import validate, ValidationError
 
 from scripts.run_timetest import check_positive_int
