+++ tests/time_tests/test_runner/utils.py
@@ -6,17 +6,22 @@
 import os
 import platform
 import sys
-from enum import Enum
-from pathlib import Path
-
 import distro
 import yaml
+import numpy as np
+
+from enum import Enum
+from pathlib import Path
 from pymongo import MongoClient
 
 # constants
 DATABASE = 'timetests'   # database name for timetests results
 DB_COLLECTIONS = ["commit", "nightly", "weekly"]
 PRODUCT_NAME = 'dldt'   # product name from build manifest
+
+# Define a range to cut outliers which are < Q1 − IQR_CUTOFF * IQR, and > Q3 + IQR_CUTOFF * IQR
+# https://en.wikipedia.org/wiki/Interquartile_range
+IQR_CUTOFF = 1.5
 
 
 def expand_env_vars(obj):
@@ -34,16 +39,14 @@ def expand_env_vars(obj):
 
 
 def upload_timetest_data(data, db_url, db_collection):
-    """ Upload timetest data to database
-    """
+    """ Upload timetest data to database."""
     client = MongoClient(db_url)
     collection = client[DATABASE][db_collection]
     collection.replace_one({'_id': data['_id']}, data, upsert=True)
 
 
 def metadata_from_manifest(manifest: Path):
-    """ Extract commit metadata from manifest
-    """
+    """ Extract commit metadata from manifest."""
     with open(manifest, 'r') as manifest_file:
         manifest = yaml.safe_load(manifest_file)
     repo_trigger = next(
@@ -58,11 +61,26 @@ def metadata_from_manifest(manifest: Path):
     }
 
 
-class UnsupportedOsError(Exception):
-    """
-    Exception for unsupported OS type
-    """
+def calculate_iqr(stats: list):
+    """IQR is calculated as the difference between the 3th and the 1th quantile of the data."""
+    q1 = np.quantile(stats, 0.25)
+    q3 = np.quantile(stats, 0.75)
+    iqr = q3 - q1
+    return iqr, q1, q3
 
+
+def filter_timetest_result(stats: dict):
+    """Identify and remove outliers from time_results."""
+    for step_name, time_results in stats.items():
+        iqr, q1, q3 = calculate_iqr(time_results)
+        cut_off = iqr * IQR_CUTOFF
+        upd_time_results = [x for x in time_results if x > q1 - cut_off or x < q3 + cut_off]
+        stats.update({step_name: upd_time_results})
+    return stats
+
+
+class UnsupportedOsError(Exception):
+    """Exception for unsupported OS type."""
     def __init__(self, *args, **kwargs):
         error_message = f'OS type "{get_os_type()}" is not currently supported'
         if args or kwargs:
@@ -72,9 +90,7 @@ class UnsupportedOsError(Exception):
 
 
 class OsType(Enum):
-    """
-    Container for supported os types
-    """
+    """Container for supported os types."""
     WINDOWS = 'Windows'
     LINUX = 'Linux'
     DARWIN = 'Darwin'
@@ -91,17 +107,17 @@ def get_os_type():
 
 
 def os_type_is_windows():
-    """Returns True if OS type is Windows. Otherwise returns False"""
+    """Returns True if OS type is Windows. Otherwise returns False."""
     return get_os_type() == OsType.WINDOWS.value
 
 
 def os_type_is_linux():
-    """Returns True if OS type is Linux. Otherwise returns False"""
+    """Returns True if OS type is Linux. Otherwise returns False."""
     return get_os_type() == OsType.LINUX.value
 
 
 def os_type_is_darwin():
-    """Returns True if OS type is Darwin. Otherwise returns False"""
+    """Returns True if OS type is Darwin. Otherwise returns False."""
     return get_os_type() == OsType.DARWIN.value
